Testing float restriction removal

+++ example_generator/app.py
@@ -93,9 +93,9 @@ def main(args):
                 logging.info('received label {}'.format(label))
                 logging.info('received filename {}'.format(infilename))
                 logging.info('downloading image')
-                image = np.array(img.getdata()).reshape(1,img.size[0], img.size[1], 3).astype('float32')
+                image = np.array(img.getdata()).reshape(1,img.size[0], img.size[1], 3)
                 logging.info('downloaded image')
-                images = np.ndarray(shape=(2,32,32,3)).astype('float32')
+                images = np.ndarray(shape=(2,32,32,3))
                 logging.info('created images storage')
                 images[0] = image
                 logging.info('assigned image to images')
